Simplify rotation using `_Numeric`
