+++ Sources/EasyImagy/ImageProtocol.swift
@@ -129,62 +129,4 @@ extension ImageProtocol {
         
         return Image(width: width, height: height, pixels: pixels)
     }
-
-    public func rotated(byDegrees angle: Int) -> Image<Pixel> {
-        precondition(angle % 90 == 0, "`angle` must be a multiple of 90: \(angle)")
-        return rotated(byRightAngleInDegrees: angle)
-    }
-
-    internal func rotated(byRightAngleInDegrees angle: Int) -> Image<Pixel> {
-        assert(angle % 90 == 0, "`angle` must be a multiple of 90: \(angle)")
-        
-        switch (angle / 90) % 4 {
-        case 0:
-            if let zelf = self as? Image<Pixel> {
-                return zelf
-            } else {
-                return Image(self)
-            }
-        case 1, -3:
-            var pixels = [Pixel]()
-            
-            for y in xRange {
-                for x in yRange.reversed() {
-                    pixels.append(self[y, x])
-                }
-            }
-            
-            return Image(width: height, height: width, pixels: pixels)
-        case 2, -2:
-            var pixels = [Pixel]()
-            
-            for y in yRange.reversed() {
-                for x in xRange.reversed() {
-                    pixels.append(self[x, y])
-                }
-            }
-            
-            return Image(width: width, height: height, pixels: pixels)
-        case 3, -1:
-            var pixels = [Pixel]()
-            
-            for y in xRange.reversed() {
-                for x in yRange {
-                    pixels.append(self[y, x])
-                }
-            }
-            
-            return Image(width: height, height: width, pixels: pixels)
-        default:
-            fatalError("Never reaches here.")
-        }
-    }
-
-    public func rotated(by angle: Double, extrapolatedBy extrapolationMethod: ExtrapolationMethod<Pixel>) -> Image<Pixel> {
-        return rotatedImageWith(angle: angle) { self[Int(round($0)), Int(round($1)), extrapolatedBy: extrapolationMethod] }
-    }
-    
-    public func rotated(byDegrees angle: Double, extrapolatedBy extrapolationMethod: ExtrapolationMethod<Pixel>) -> Image<Pixel> {
-        return rotated(by: angle / 180.0 * .pi, extrapolatedBy: extrapolationMethod)
-    }
 }
