--- Sources/EasyImagy/ImageProtocolTyped.swift.gyb
@@ -1,83 +1,0 @@
-% channel_types = [
-%   'UInt8',
-%   'UInt16',
-%   'UInt32',
-%   'Int',
-%   'Float',
-%   'Double',
-% ]
-% types = [f'RGBA<{t}>' for t in channel_types] + channel_types
-%
-import Foundation
-% for i, type in enumerate(types):
-
-extension ImageProtocol where Element == ${type} { // Rotation
-    public func rotated(byDegrees angle: Int) -> Image<Pixel> {
-        if angle % 90 == 0 {
-            return rotated(byRightAngleInDegrees: angle)
-        } else {
-            return rotated(byDegrees: Double(angle))
-        }
-    }
-
-    public func rotated(by angle: Double) -> Image<Pixel> {
-        return rotatedImageWith(angle: angle) { self[$0, $1, interpolatedBy: .bilinear, extrapolatedBy: .filling(.selfZero)] }
-    }
-    
-    public func rotated(byDegrees angle: Double) -> Image<Pixel> {
-        return rotated(by: angle / 180.0 * .pi)
-    }
-
-    public func rotated(by angle: Double, extrapolatedBy extrapolationMethod: ExtrapolationMethod<Pixel>) -> Image<Pixel> {
-        return rotatedImageWith(angle: angle) { self[$0, $1, interpolatedBy: .bilinear, extrapolatedBy: extrapolationMethod] }
-    }
-    
-    public func rotated(byDegrees angle: Double, extrapolatedBy extrapolationMethod: ExtrapolationMethod<Pixel>) -> Image<Pixel> {
-        return rotated(by: angle / 180.0 * .pi, extrapolatedBy: extrapolationMethod)
-    }
-
-    public func rotated(by angle: Double, interpolatedBy interpolationMethod: InterpolationMethod, extrapolatedBy extrapolationMethod: ExtrapolationMethod<Pixel>) -> Image<Pixel> {
-        return rotatedImageWith(angle: angle) { self[$0, $1, interpolatedBy: interpolationMethod, extrapolatedBy: extrapolationMethod] }
-    }
-    
-    public func rotated(byDegrees angle: Double, interpolatedBy interpolationMethod: InterpolationMethod, extrapolatedBy extrapolationMethod: ExtrapolationMethod<Pixel>) -> Image<Pixel> {
-        return rotated(by: angle / 180.0 * .pi, interpolatedBy: interpolationMethod, extrapolatedBy: extrapolationMethod)
-    }
-}
-% end
-
-extension ImageProtocol where Element == RGBA<Bool> { // Rotation
-    public func rotated(byDegrees angle: Int) -> Image<Pixel> {
-        if angle % 90 == 0 {
-            return rotated(byRightAngleInDegrees: angle)
-        } else {
-            return rotated(byDegrees: Double(angle))
-        }
-    }
-
-    public func rotated(by angle: Double) -> Image<Pixel> {
-        return rotatedImageWith(angle: angle) { self[Int(round($0)), Int(round($1)), extrapolatedBy: .filling(RGBA<Bool>(red: false, green: false, blue: false, alpha: false))] }
-    }
-    
-    public func rotated(byDegrees angle: Double) -> Image<Pixel> {
-        return rotated(by: angle / 180.0 * .pi)
-    }
-}
-
-extension ImageProtocol where Element == Bool { // Rotation
-    public func rotated(byDegrees angle: Int) -> Image<Pixel> {
-        if angle % 90 == 0 {
-            return rotated(byRightAngleInDegrees: angle)
-        } else {
-            return rotated(byDegrees: Double(angle))
-        }
-    }
-
-    public func rotated(by angle: Double) -> Image<Pixel> {
-        return rotatedImageWith(angle: angle) { self[Int(round($0)), Int(round($1)), extrapolatedBy: .filling(false)] }
-    }
-    
-    public func rotated(byDegrees angle: Double) -> Image<Pixel> {
-        return rotated(by: angle / 180.0 * .pi)
-    }
-}
